Adjust default rate-limits to be wide-open
This commit adjusts our default tower limits to be wide open, effectively with
no rate limit. It is important for endpoints that dictate rate limits that we
have some mechanism in place for this but, overall, most endpoints do not set
explicit limits so why should we throttle ourselves?

Follow-up work will trawl through each sink and adjust where
higher-than-previous defaults have been set but are still quite low.

REF #8470
