+++ docs/cue/reference/components/sinks/new_relic_logs.cue
@@ -36,7 +36,6 @@ components: sinks: new_relic_logs: {
 				enabled:                    true
 				concurrency:                100
 				rate_limit_duration_secs:   1
-				rate_limit_num:             100
 				retry_initial_backoff_secs: 1
 				retry_max_duration_secs:    10
 				timeout_secs:               30
